+++ dev/sparktestsupport/modules.py
@@ -710,6 +710,9 @@ pyspark_pandas = Module(
         "pyspark.pandas.tests.data_type_ops.test_null_ops",
         "pyspark.pandas.tests.data_type_ops.test_num_ops",
         "pyspark.pandas.tests.data_type_ops.test_num_arithmetic",
+        "pyspark.pandas.tests.data_type_ops.test_num_mod",
+        "pyspark.pandas.tests.data_type_ops.test_num_mul_div",
+        "pyspark.pandas.tests.data_type_ops.test_num_pow",
         "pyspark.pandas.tests.data_type_ops.test_num_reverse",
         "pyspark.pandas.tests.data_type_ops.test_string_ops",
         "pyspark.pandas.tests.data_type_ops.test_udt_ops",
@@ -1201,6 +1204,9 @@ pyspark_pandas_connect_part1 = Module(
         "pyspark.pandas.tests.connect.series.test_parity_sort",
         "pyspark.pandas.tests.connect.series.test_parity_stat",
         "pyspark.pandas.tests.connect.data_type_ops.test_parity_num_arithmetic",
+        "pyspark.pandas.tests.connect.data_type_ops.test_parity_num_mod",
+        "pyspark.pandas.tests.connect.data_type_ops.test_parity_num_mul_div",
+        "pyspark.pandas.tests.connect.data_type_ops.test_parity_num_pow",
         "pyspark.pandas.tests.connect.reshape.test_parity_get_dummies",
         "pyspark.pandas.tests.connect.reshape.test_parity_get_dummies_kwargs",
         "pyspark.pandas.tests.connect.reshape.test_parity_get_dummies_multiindex",
